feat(forum_conversation): renaming www app
