+++ lacommunaute/www/forum_conversation_views/tests.py
@@ -9,7 +9,7 @@ from lacommunaute.forum_conversation.forms import PostForm
 from lacommunaute.forum_conversation.models import Topic
 from lacommunaute.forum_upvote.factories import UpVoteFactory
 from lacommunaute.users.factories import UserFactory
-from lacommunaute.www.forum_conversation.views import PostListView
+from lacommunaute.www.forum_conversation_views.views import PostListView
 
 
 TopicReadTrack = get_model("forum_tracking", "TopicReadTrack")
